Refactored all of app.py except budgets and cashflows
